config publish locations - partial
