+++ src/stubber/publish/publish_stubs.py
@@ -31,7 +31,6 @@ import sys
 from pathlib import Path
 from typing import Dict, List, Tuple, Union
 
-import pytest
 from loguru import logger as log
 from packaging.version import parse
 from pysondb import PysonDB
@@ -49,17 +48,6 @@ COMBO_STUBS = ALL_TYPES[0]
 DOC_STUBS = ALL_TYPES[1]
 CORE_STUBS = ALL_TYPES[2]
 
-
-def get_database(root_path: Union[Path, str], production: bool = False) -> PysonDB:
-    """
-    Open the json database at the given path.
-
-    The database should be located in a subfolder `/publish` of the root path.
-    The database name is determined by the production flag as `package_data[_test].jsondb`
-    """
-    root_path = Path(root_path)
-    db_path = root_path / f"package_data{'' if production else '_test'}.jsondb"
-    return PysonDB(db_path.as_posix())
 
 
 def package_name(port: str = "", board: str = "", pkg_type=COMBO_STUBS, family="micropython") -> str:
